Rename fonduerconfig.py to fonduer_subclasses.py and import mention_classes & candidate_classes in fonduer_model.py
